+++ src/main/java/com/web2/movelcontrol/Model/Pedido.java
@@ -29,17 +29,17 @@ public class Pedido {
     public Pedido() {
     }
 
-    public Pedido(Integer id, Date data_pedido, Orcamento orcamento) {
+    public Pedido(Long id, Date data_pedido, Orcamento orcamento) {
         this.id = id;
         this.data_pedido = data_pedido;
         this.orcamento = orcamento;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
