+++ src/main/java/com/web2/movelcontrol/Model/Pessoa.java
@@ -20,7 +20,7 @@ public abstract class Pessoa {
     public Pessoa() {
     }
 
-    public Pessoa(Integer id, String nome, String telefone, String email, String endereco) {
+    public Pessoa(Long id, String nome, String telefone, String email, String endereco) {
         this.id = id;
         this.nome = nome;
         this.telefone = telefone;
@@ -28,11 +28,11 @@ public abstract class Pessoa {
         this.endereco = endereco;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
